rename sign/api to sign/schemes
